test(runtime): extract remote-agent tests into config/remote_agent_tests module (#913)
